bidouilles
J'ai modifié des trucs pour optimiser le code ! Je n'ai toujours pas réussi à faire apparaître le score dans le tableau.
Mais maintenant si le joueur ne rentre pas de pseudo, un lui ai attribué automatiquement.

+++ SpaceInvador/src/Ligne.java
@@ -56,11 +56,11 @@ public class Ligne {
 //	}
 
 	
-	public boolean testerPlancher(int hauteur)
+	public boolean testerPlancher(double hauteur)
 	{
 		ListIterator<Alien> iterAlien = listeAlien.listIterator();
 		while(iterAlien.hasNext())
-			if(iterAlien.next().getY()+Alien.COTE > hauteur)
+			if(iterAlien.next().getY() + Alien.COTE > hauteur)
 				return true;
 		return false;
 	}
